requirements.txt
# Flask framework for building the web API
flask

# MySQL connector for Python
mysql-connector-python

# For loading environment variables from a .env file
python-dotenv

# WSGI server
waitress